Add content for 'articles' and 'projects' in multiple languages
Introduced new sections 'articles' and 'projects' in both English and German language configurations. This was done in the 'en-US.json' and 'de-DE.json' files, respectively. For both languages, sections were also implemented on the webpage using Vue.js. This enables the user to read about various articles and explore different projects in their preferred languages for a more personalized and interactive experience.