Refactor code to improve article component and layout
The article.vue layout has been deleted and replaced by a new article component that is used within pages to wrap content. This provides more flexibility in defining the layout of individual pages. Alongside this, changes are made in several other components to enhance user interface, such as improved hovering effects, changes in colour scheme and adjustments to the display of pivotal elements based on screen size. Several unused or redundant code blocks are also removed.