Refactor layout component and resources markdown file
The layout component was updated by implementing a new Article component instead of the existing article.vue layout. This makes it easier to define individual page layouts. The resources markdown file has been significantly trimmed down and reorganized for improved presentation and readability. In addition, an unnecessary About page was removed, the resources.vue file underwent modifications, and a new test file was added to the projects directory.