Update resources.md, package.json, and README.md files
The position of the Nuxt/ColorMode link has been rearranged in the resources.md file and description updated for a banner item. A new "start" script has been added to the package.json file. Detailed setup instructions have been added to the README.md for easier understanding and implementation for users.